Factor out parse-compiled-messages.ts.
