+++ frontend/localebuilder/cli.ts
@@ -1,102 +1,7 @@
 import fs from "fs";
 import PO from "pofile";
-import * as babel from "@babel/core";
 import generate from "@babel/generator";
-
-class LinguiMessageCatalog {
-  constructor(
-    readonly languageData: babel.types.ObjectExpression,
-    readonly messages: Map<string, babel.types.Node>
-  ) {}
-}
-
-class LinguiMessageCatalogParser {
-  languageData: babel.types.ObjectExpression | null = null;
-  messages: Map<string, babel.types.Node> = new Map();
-  t: typeof babel.types = babel.types;
-
-  private constructor(js: string) {
-    const file = babel.parseSync(js);
-    if (file && file.type === "File") {
-      this.visitProgram(file.program);
-    }
-  }
-
-  static parse(js: string): LinguiMessageCatalog {
-    const lmcp = new LinguiMessageCatalogParser(js);
-    const { languageData, messages } = lmcp;
-    if (!languageData) {
-      throw new Error("languageData not found!");
-    }
-    return new LinguiMessageCatalog(languageData, messages);
-  }
-
-  visitProgram(program: babel.types.Program) {
-    for (let statement of program.body) {
-      this.visitBodyStatement(statement);
-    }
-  }
-
-  visitBodyStatement(statement: babel.types.Statement) {
-    const { t } = this;
-    if (
-      t.isExpressionStatement(statement) &&
-      t.isAssignmentExpression(statement.expression) &&
-      t.isObjectExpression(statement.expression.right)
-    ) {
-      this.visitModuleExports(statement.expression.right);
-    }
-  }
-
-  visitModuleExports(objectExpression: babel.types.ObjectExpression) {
-    const { t } = this;
-    for (let property of objectExpression.properties) {
-      if (t.isProperty(property)) {
-        this.visitModuleExport(property);
-      }
-    }
-  }
-
-  visitModuleExport(property: babel.types.Property) {
-    const { t } = this;
-    if (t.isIdentifier(property.key)) {
-      const { name } = property.key;
-      const { value } = property;
-      if (name === "languageData") {
-        this.visitLanguageDataExport(value);
-      } else if (name === "messages") {
-        this.visitMessagesExport(value);
-      }
-    }
-  }
-
-  visitLanguageDataExport(value: babel.types.Node | null) {
-    if (this.t.isObjectExpression(value)) {
-      this.languageData = value;
-    }
-  }
-
-  visitMessagesExport(value: babel.types.Node | null) {
-    const { t } = this;
-    if (t.isObjectExpression(value)) {
-      for (let property of value.properties) {
-        this.visitMessageProperty(property);
-      }
-    }
-  }
-
-  visitMessageProperty(
-    property:
-      | babel.types.ObjectMethod
-      | babel.types.ObjectProperty
-      | babel.types.SpreadElement
-  ) {
-    const { t } = this;
-    if (t.isProperty(property) && t.isStringLiteral(property.key)) {
-      this.messages.set(property.key.value, property.value);
-    }
-  }
-}
+import { parseCompiledMessages } from "./parse-compiled-messages";
 
 function getFilepath(reference: string) {
   const parts = reference.split(":");
@@ -114,7 +19,7 @@ export function run() {
   const messagesJs = fs.readFileSync("locales/es/messages.js", {
     encoding: "utf-8",
   });
-  const mc = LinguiMessageCatalogParser.parse(messagesJs);
+  const mc = parseCompiledMessages(messagesJs);
   const src = generate(mc.languageData);
   console.log("languageData is", src.code);
   for (let [message, value] of mc.messages) {
